Main de prueba, agregado de readCity. Problema del addrec?

+++ cityADT.h
@@ -1,6 +1,7 @@
 #ifndef __CITYADT_H
 #define __CITYADT_H
 #include <string.h>
+#include <stdlib.h>
 
 // Valor default de ERROR sera -1 para todo el archivo
 
@@ -38,18 +39,18 @@ int addTree(cityADT c, char * neigh, char *tree);
 ** el vector esta ordenado de manera alfabetica por barrio.
 ** retorna NULL y dim en -1 en caso de ERROR, dim en 0 si esta vacio
 */
-char ** mostPopularTree(cityADT c , int * dim);
+char ** mostPopularTree(cityADT c , long int * dim);
 
 /*
 **devuelve barrios en orden alfabético
 ** dim parametro de entrada y salida con la cantidad de barrios
 ** retorna NULL y dim en -1 en caso de ERROR, dim en 0 si esta vacio
 */
-char ** showAllNeigh( cityADT c , int *dim );
+char ** showAllNeigh( cityADT c , long int *dim );
 
 /*devuelve cant de arboles por persona en cada barrio (alfabeticamente)
 ** retorna NULL y dim en -1 en caso de ERROR, dim en 0 si esta vacio
 */
-double* treesPerPerson ( cityADT c , int *dim);
+double* treesPerPerson ( cityADT c , long int *dim);
 
 #endif
